[change] Separate polyfills from main chunk
