+++ client/polyfills.js
@@ -1,0 +1,2 @@
+import "core-js/stable";
+import "regenerator-runtime/runtime";
